+++ frontend/lib/pages/checkpoints.dart
@@ -74,7 +74,7 @@ class CheckPointsModel extends DropDownModel<CheckPoint, CheckCategory>
             checkpoint['PjNr'] = currentData.pjNr;
             checkpoint['E1'] = currentData.index;
             checkpoint['E2'] = -1;
-            await Backend()
+            await API()
                 .setNew(CheckPoint.fromJson(checkpoint), caller: currentData);
             notifyListeners();
           },
